Install Terraform before running tests
We're seeing occasional test failures caused by the `terraform` binary being modified while we're trying to execute it as part of the acceptance tests (see this related issue: https://github.com/hashicorp/terraform-plugin-sdk/issues/629).

The suggested solution in the Terraform plugin developer docs is to install a version of Terraform before running `go test`, which should stop the tests from trying to download Terraform themselves.

I had to switch from using the Go container to the `setup-go` action because the `setup-terraform` action relies on `unzip`, which wasn't in the Go container.