Better error message when importing missing stack (#406)
The terraform import handling for stacks was reusing the `resourceStackRead()` function, which sets the ID to an empty string if no stack is found. To solve this I've pulled out query for a stack and updating the state based on a stack struct into separate functions, and I've updated the import handling to detect when the stack can't be found an output a decent error message.

I pulled the `PopulateStack()` function into the structs package to follow the pattern we've taken for the scheduled task structs.