Use correct retry logic when handling AWS roles and their attatchments
We recently updated the error message returned from our server to help clarify to users of the spacelift
UI what actions they need to take to resolve their issue. However we missed updating the retry loop here specifically
checking for the text of the old error.

I will be making a comment internally too to ensure that any future changes to those error messages are also reflected here
in the provider.

Signed-off-by: James Humphries <james@james-humphries.co.uk>