fix: allow to manually run external contributors workflows
This changed the way we run the test workflow that requires secrets.

Actually we should close contributors PRs and reopen ourselves to be
able to run tests. That is a massive PITA.

Dependabot PRs are also broken.

This changes how we run this test workflow by using pull_request_target
and then manually checking for permissions. We also always allow
dependabot to run.

So for an external contributor, our workflow would be:

- external contributor fork the project and create a new PR targetting
main
- the test workflow run and fails
- one internal maintainer review the PR
- internal maintainers can retry the failed run if the PR looks safe
- all green, we can merge